[idleaction] Patch xprintidle to not use the creepy workaround anymore
This used to work around a bug in the xserver where it would always
reset the idle time whenever the screen goes to standby or suspend.
This does not happen anymore on a recent system, so xprintidle would
erroneously add the timeout values to the idletime, resulting in
incorrect high results when the screen is turned off.