[beamergui] Fix mode setter (once again)
First, reading ip-based config from beamer.conf was goofed,
which isn't the preferred way of doing things anyways and
then if nothing is found in the config, the fallback path
that sets both outputs to 1024x768 wasn't taken.
...which brings me to the second point that 1024x768 is quite
ancient, so let's fall back to the screen's preferred resolution
instead, and if that fails too, use 1280x720 which should
be an ok compromise and not make any halfway modern projector
explode.