[idleaction] Set wakeup alarm two minutes early to prevent race
If the wakeup is scheduled exactly at the time a scheduled
shutdown is, the cronjob for the shutdown would not run, as
the client is still about to wake up at that time.
If we wake up two minutes early, the cron daemon should have
enough time to fully wake up, notice the clock jump, and then
finally trigger the scheduled shutdown.