[pam-slx-plug] Add get_username script to lookup caps
Changing the capitalization of usernames in the PAM stack is a bad idea
and messes everything up. Add a helper script that can loop up a user
name and return the proper capitalization. This script should be called
in the lightdm greeter. exec_auth now refuses login if the username
returned by LDAP has a different capitalization.

Fixes #3503