[slx-issue] use baseboard dmi info as fallback
some models like Intel's NUC5i5RYB do not provide
system-manufacturer/system-product-name in the DMI information -
that information is stored in baseboard-* instead. So we use that as a
fallback in case system-* was empty. Use '<unknown>' as ultimate
fallback.

This kinda works around an issue in the final loop filling in spaces: if
a line's value is long enough (e.g. 34 whitespaces like a untrimmed
'dmidecode -s' output...) the space variable would never be set. Thus,
${space:0:-2} would fail unless ${#space} is at least 2 and abort the
creation of /etc/issue altogether.