[pam-slx-plug] Add auth-final-exec hook
On successful authentication, run everything
in dir /opt/openslx/pam/hooks/auth-final-exec.d
This applies no matter which authentication module
succeeded, contrary to the old pam_script_auth.d directory.
Note that the password is NOT exposed in this hook,
and it is only run if the pam stack is executing in root
context.