[kiosk-*] Support SLX_BROWSER, add precheck for browser binary
SLX_BROWSER is is space separated list of keywords. They are globbed
within the kiosk-launch.d directory, so specifying
SLX_BROWSER='chromium firefox'
would first look for any file named *chromium* and source it, and if
that fails for any file named *firefox*.
If everything fails or the variable is empty, any file in that directory
will be sourced.