[vbox-src] add vboxusers group on build-time
this makes sure the vboxusers group is available during boot and that the
vbox's udev rules will not trigger before the group is created. If they
did, /dev/vboxusb would be owned by root:root, which would prevent users
from the group vboxusers to access it and would lead to the first user
session to not see USB devices.