[beamergui] Support SLX vars for manual screen config
SLX_FORCE_RESOLUTION is a list of space separated resolutions. It will
be applied to all outputs in alphabetical order. If there are more
outputs than resolutions in this list, it will "loop around", starting
to assign resolutions from the beginning of the list.

SLX_RESOLUTION_MAPPING can be given in addition to the above. In this
case, the resolutions won't be assigned in alphabetical order. The
variable is supposed to contain a list of space separated assignments
of the form <output>=<index>, for example
HDMI-2=0 DP-3=0 HDMI-1=1
would assign HDMI-2 and DP-3 to the first resolution of SLX_FORCE_RESOLUTION
and HDMI-1 to the second resolution, making the first two outputs a
clone of each other, and HDMI-1 extend to the right.