[remote-access] Disable ZaphodHead on intel
This can cause the X session to become unresponsive under certain,
currently unknown circumstances. It seems to involve VMware 15.5.x vs.
15.1.x previously, and happens after the xscreensaver tried to start
once (and fails).
This hasn't been reproduced everywere and remains a mystery, but until
this feature works reliably, we have to forgo the warning messages on
the physically connected screen and let it enter standby instead.