[etherwake/idleaction] Try harder to enable WOL
It was reported that WOL is disabled at runtime on some machines, even
though there is a udev rule in place that enables it right when the
interfaces get registered. We can only assume this is a race condition
with adapter initialization, but worst case something else is running
that is interfering with the WOL setting. Now we forcefully re-enable
WOL at boot, every hour, and before standby, for good measure.