[kiosk-chromium] if interactive mode and user was active, reload -> exit
If a user was active (measured by idle timer) and chromium is running in
interactive mode, i.e. with full UI and tabs, exit the session if the
reload timeout is reached. This is a security measure, since we assume
that the browser in interactive mode enables the user to log into
websites etc., and might not log out everywhere when leaving.