Implement blocking network.target and network-online.target
Add two services that will delay reaching the according targets, until
conditions we consider reasonable are reached.
For network.target that is finishing DHCP for the primary interface, for
network-online.target we want to see a default gateway or a proxy server
configured.
Both have a timeout of about 10 seconds for now; this might be increased
later on, if necessary.