[system-check/hardware-stats] Refactor report stats gather split stuff
Reporting to server is now a separate service, as it slows down
system-check for no reason via hooks.d.
Also add a couple comments and simplifications to system-check main
script. We now redirect to the tags file within the main script instead
of letting each hook individually open the tags file and write to it
concurrently, which seems like a bad idea in retrospect.