[hardware-stats] Fix detection of vmx/svm support
Newer Linux kernel versions hide the vmx/svm flag from /proc/cpuinfo if
you disable the feature in the BIOS. We wrongfully reported this as
UNSUPPORTED then instead of DISABLED as intended. Read the according
MSRs unconditionally now, and only validate the flag is present if the
MSR values line up with what we expect for the feature to be enabled.