[run-virt] Add forced logout option to pvs toolbar
If a VM session freezes hard, there is no trivial way to end the
session. For normal sessions, this is not the end of the world, since
the user can simply reboot the machine. For remote sessions however, you
need to rely on the idle timeout of abandoned sessions, which is all but
obvious to the user.

Closes #3911