[inc/chroot] Fix minor issues, work-around tmfs umount weirdness
It seems if we mount a tmfs after establishing a bind-mount of /, we get
an additional entry in /proc/mounts, and unmounting either one
will remove the other one too. This breaks keeping the upperdir
of a chroot in case of failure, making debugging hard. Change the
order: mount the tmpfs first, then the bind-mount of /. This avoids
the additional entry and keeps the upperdir around.