[virt-manager] Set prefix to /usr/local to avoid broken build
For some reason, virt-manager always installs to /usr/local,
no matter what you set --prefix to. However, in that case it will
"helpfully" write the wrong path to the virt-manager startup wrapper,
which then points to nowhere. So set the prefix to where it stubbornly
installs anyways to fix this issue.