[dhcpc-busybox] Don't wait for dhcp on br0 anymore
We did this to avoid issues if the hostname changes after Xorg
started up. But sometimes, DHCP is really slow, and we delay the
startup of the login screen by a good 10 seconds. Avoid this
by instead just touching the flag file that prevents the udhcpc
trigger script from changing the hostname. Running with a
potentially wrong hostname should be an acceptable tradeoff.