[dnbd3-proxy-mode] Dynamic deletion delay based on RTC
If the proxy is being started up mid-day, it might be undesired to have
it wait for 18 hours before deleting old images. This might mean we
don't have a good idea which image has been unused longest, but it's
better to delete *some* unused image than just not replicating any
images at all.