[qemu] Passthrough: Find free PCI slot on primary PCI bus
Explicitly assign a slot number on the primary PCI bus to the VGA card,
otherwise the two functions of the GPU (gfx and sound) might end up
in different slots or even on different buses, which might cause issues.
We now put all parts of the GPU in one slot and copy the original
function IDs.