[system-check/hardware-stats] Loosen dependencies
For some reason, system-check is ordered before hardware-report, but I
can't figure out why. system-check merely generates the warnings to
display on the login screen, and hardware-report sends hw data to the
boot server. I don't see anything in hardware-report actually depending
on anything system-check does.

There are some other minor changes, like moving the generation of the
json report from hardware-report to hardware-stats, and putting the
according python script into a directory named hardware-stats instead of
system-check, which might imply it belongs to that module which is wrong.