[bwlp-stage4-tweaks] Add service to grow rootfs to blockdev size
This was previously done in initrd, but it turns out this was happening
rather late and blocked the switchroot for a couple seconds. Let's do it
in stage 4 instead as early as possible, and order it before
gather-hw-info.service, as that service will determine the rootfs free
space and needs to see the properly resized rootfs.