[hardware-stats] Use optimistic space calc to drop grow-rootfs dep
grow-rootfs might take a while. Use the newly introduced --optimistic
option to fs_path_space so it will consider the size of the underlying
device when reporting the size and available space, instead of what the
fileystem itself reports. This is so that when we're still resizing the
partition (which can take about 10 seconds on some systems), we will
already get a somewhat correct estimate. Since this is just for display
in slx-admin, nothing crocial depends on it.